src/main/java/tobyspring/helloboot/HelloController.java
package tobyspring.helloboot;

import org.springframework.context.ApplicationContext;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Objects;

//클래스 레벨의 매핑을 찾은 후에 메쏘드 레벨을 찾음.
@RestController
public class HelloController {
    //변수로 먼저 저장 하고 사용
    private final HelloService helloService;

    public HelloController(HelloService helloService) {
        this.helloService = helloService;
    }

    @GetMapping("/hello")
    public String hello(String name) {
        if (name == null || name.trim().length() == 0) throw new IllegalArgumentException();

        return helloService.sayHello(name);
    }

}
